Refactor common method related to creating random names
